refactor graphql subscriptions
