+++ Core/Core/Api/GraphQL/Client.GraphqlCleintOperations/Client.ServerOperations.cs
